quitando archivos de merge
