+++ giveFeedBack.html
@@ -26,5 +26,5 @@
     </div>
     
 </body>
-<script src="./js/giveFeedBack.js"></script>
+<script type="module" src="./js/giveFeedBack.js"></script>
 </html>
