--- singIn.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Sing in</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
-</head>
-
-<body>
-    <!-- <form class="m-5" method="" action="" onsubmit="event.preventDefault()">  -->
-    <form class="m-5" onsubmit="event.preventDefault()" id="form"> 
-
-        <div class="form-group">
-            <label for="inputName">Name</label>
-            <input type="text" class="form-control" id="inputName">
-          </div>
-
-        <div class="form-group">
-          <label for="inputEmail">Email address</label>
-          <input type="email" class="form-control" id="inputEmail">
-        </div>
-
-        <div class="form-group">
-          <label for="inputPassword">Password</label>
-          <input type="password" class="form-control" id="inputPassword">
-          <small id="passwordHelp" class="form-text text-muted">For security the password must be greater than 8 digitos.</small>
-        </div>
-
-        <button onclick="button()" type="submit" class="btn btn-primary">Submit</button>
-    </form>
-    
-    <script src="./singIn.js "></script>
-</body>
-</html>
